improve handling of validation on field blur
When a field with an error is focused, and a cancel button was clicked, the button wouldn’t work, because the field losing focus would trigger an error message to show up, moving the button. Now instead putting an action on blur event, the form keeps track of which fields have been focused, and when a field has been focused but another field subsequently gets focus, the first is considered blurred.  This means validation errors only pop up when moving focus to another field, and not a cancel button or any other element.